requirements.txt
reflected-ffi==0.0.3
flatted-view==0.0.1
next-resolver==0.0.2
websockets
fastapi
argparse
uvicorn
build
twine
